+++ locales/fr/LC_MESSAGES/volto.po
@@ -173,6 +173,10 @@ msgstr "Cacher"
 msgid "Hide all content types"
 msgstr "Masquer tous les types de contenu"
 
+#: components/ItaliaTheme/Breadcrumbs/Breadcrumbs
+msgid "Home"
+msgstr ""
+
 #: components/ItaliaTheme/Blocks/NumbersBlock/Edit
 msgid "Icons placeholder"
 msgstr ""
