+++ locales/it/LC_MESSAGES/volto.po
@@ -156,6 +156,10 @@ msgstr "Nascondi"
 msgid "Hide all content types"
 msgstr ""
 
+#: components/ItaliaTheme/Breadcrumbs/Breadcrumbs
+msgid "Home"
+msgstr ""
+
 #: components/ItaliaTheme/Blocks/NumbersBlock/Edit
 msgid "Icons placeholder"
 msgstr "Seleziona le icone dalla barra a lato"
