+++ locales/ja/LC_MESSAGES/volto.po
@@ -163,6 +163,10 @@ msgstr ""
 msgid "Hide all content types"
 msgstr ""
 
+#: components/ItaliaTheme/Breadcrumbs/Breadcrumbs
+msgid "Home"
+msgstr ""
+
 #: components/ItaliaTheme/Blocks/NumbersBlock/Edit
 msgid "Icons placeholder"
 msgstr ""
