+++ locales/volto.pot
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: Plone\n"
-"POT-Creation-Date: 2021-07-19T10:11:23.258Z\n"
+"POT-Creation-Date: 2021-07-19T14:20:18.864Z\n"
 "Last-Translator: Plone i18n <plone-i18n@lists.sourceforge.net>\n"
 "Language-Team: Plone i18n <plone-i18n@lists.sourceforge.net>\n"
 "MIME-Version: 1.0\n"
@@ -189,6 +189,11 @@ msgstr ""
 #: components/ItaliaTheme/Search/SearchCTs
 # defaultMessage: Nascondi tutti i tipi di contenuto
 msgid "Hide all content types"
+msgstr ""
+
+#: components/ItaliaTheme/Breadcrumbs/Breadcrumbs
+# defaultMessage: Home
+msgid "Home"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/NumbersBlock/Edit
